Use `decimal.Decimal` exclusively in spreadsheet formulas
The half-and-half use before was causing normal math operations to get
screwed up, and for complex spreadsheet formulas to return the incorrect
result.

Added tests for these cases to prevent regressions. In future it'd be
good if we could eliminate `float` entirely so that even normal math
operations don't suffer from issues with floating point precision.